Move FileUpload component into Predict folder.
